Patreon Supporters Displays (#102)
Added custom file input for plugin "settings". We pass the public information down via a hidden element in the HTML
![image](https://github.com/user-attachments/assets/12d14b5e-ff5a-4aa3-8459-8b13311b8824)

We then use that HTML to parse out the public data and format it into the thank you message. One designed for the hompeage:
![image](https://github.com/user-attachments/assets/c374aff7-be35-4c50-a659-65c73bbee9e5)

And one designed for articles:
![image](https://github.com/user-attachments/assets/36021d2e-5fe6-4bc6-8a54-20b745ba10ef)

Both are mobile compatible and can be placed by creating a div with the ids `patreon-supporters-all` or `patreon-supporters-article`

TODOs:
- [x] Light Mode compatability
- [x] Uploading after previous one was uploaded fails always, fix this if simple enough