Magazine Advertisement (#115)
Re-purposing the Patreon Pop-Up to notify users of the new Magazine Issue released by Tank Encyclopedia.

The behavior of the popup is the same in that when the user presses the X button it will close for a week. However, the popup appears upon a first-time user's second page visit since it's more time sensitive than the Patreon prompt. To be deployed this Sunday.

Tested in Edge, Chrome, and Firefox minimally, as the majority of this component has already been in production.

* implement magazine notification

* GHA_BUILD: Minimize JS Versions

* GHA_BUILD: Minimize CSS Versions

* GHA_BUILD: Increment Plugin Version to 2.1.43

* GHA_BUILD: Zip Plugin

---------

Co-authored-by: GitHub Actions <actions@github.com>