T.E. Magazine Issue 11 Native Advertisement (#126)
The new magazine is ready to go, so must the ad be. Also changed the behavior slightly so that returning users (visitors that have the old visited cookie present) get the new ad sent out immediately, but brand-new visitors (ones that don't have the old visited cookie) don't get the ad immediately. The ad is still clearable for 1 week at a time.

- [x] Double check with stakeholders that the wording and look is acceptable

<img width="295" height="320" alt="image" src="https://github.com/user-attachments/assets/2b170943-9332-4ccf-ba15-ad77733b0ccb" />