Magazine Ad 11: Hotfix (#127)
Now that we have possibly completely `undefined` cookies, we need to add a check to ensure that it exists before trying to get the value out of it. This was ok before, but now that we added an `||` condition, we have a problem. Probably don't need the `||` condition anymore, but we might as well check it as a short-circuit.